+++ tests/test_ft017th.py
@@ -18,6 +18,7 @@
 import datetime
 import logging
 import queue
+import time
 import unittest.mock
 
 import cc1101
@@ -109,67 +110,67 @@ def test__parse_transmission_repeats_dont_match(signal):
         )
 
 
-def test__receive_packet():
-    with unittest.mock.patch("cc1101.CC1101"):
-        sensor = wireless_sensor.FT017TH()
-    with unittest.mock.patch("time.sleep") as sleep_mock:
-        sensor._transceiver.get_marc_state.side_effect = (
-            lambda: cc1101.MainRadioControlStateMachineState.RX
-            if sum(a[0] for a, _ in sleep_mock.call_args_list) < 16
-            else cc1101.MainRadioControlStateMachineState.IDLE
-        )
-        sensor._transceiver._get_received_packet.side_effect = (
-            lambda: "fail"
-            if sum(a[0] for a, _ in sleep_mock.call_args_list) < 16
-            else "dummy"
-        )
-        packet = sensor._receive_packet()
-    sensor._transceiver._enable_receive_mode.assert_called_once_with()  # pylint: disable=no-member; false positive
-    assert sleep_mock.call_args_list == [
-        unittest.mock.call(0.05),
-        unittest.mock.call(8.0),
-        unittest.mock.call(8.0),
+@pytest.mark.parametrize("gdo0_gpio_line_name", (b"GPIO24", b"GPIO25"))
+@pytest.mark.parametrize("timeout_seconds", (21, 42))
+def test__receive_measurement_timeout(caplog, gdo0_gpio_line_name, timeout_seconds):
+    sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=gdo0_gpio_line_name)
+    with unittest.mock.patch.object(
+        sensor._transceiver, "_wait_for_packet", return_value=None
+    ) as wait_for_packet_mock, caplog.at_level(logging.DEBUG):
+        assert sensor._receive_measurement(timeout_seconds=timeout_seconds) is None
+    wait_for_packet_mock.assert_called_once_with(
+        gdo0_gpio_line_name=gdo0_gpio_line_name, timeout_seconds=timeout_seconds
+    )
+    assert caplog.record_tuples == [
+        ("wireless_sensor", logging.DEBUG, "timeout or fetching packet failed")
     ]
-    assert packet == "dummy"
 
 
 def test__receive_measurement_unexpected_length():
-    sensor = wireless_sensor.FT017TH()
+    sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     with unittest.mock.patch.object(
-        sensor,
-        "_receive_packet",
+        sensor._transceiver,
+        "_wait_for_packet",
         return_value=cc1101._ReceivedPacket(
-            data=b"\xff" * 17,
+            payload=b"\xff" * 17,
             rssi_index=0,
             checksum_valid=True,
             link_quality_indicator=0,
         ),
     ):
         with pytest.raises(wireless_sensor._UnexpectedPacketLengthError):
-            sensor._receive_measurement()
+            sensor._receive_measurement(timeout_seconds=21)
 
 
 def test_receive():
     with unittest.mock.patch("cc1101.CC1101"):
-        sensor = wireless_sensor.FT017TH()
-    measurement_iter = sensor.receive()
+        sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
+    measurement_iter = sensor.receive(timeout_seconds=21)
     packet = cc1101._ReceivedPacket(
-        data=b"\0" * 23, rssi_index=0, checksum_valid=True, link_quality_indicator=0
+        payload=b"\0" * 23, rssi_index=0, checksum_valid=True, link_quality_indicator=0
     )
-    with unittest.mock.patch.object(sensor, "_receive_packet") as receive_packet_mock:
+    with unittest.mock.patch.object(
+        sensor._transceiver, "_wait_for_packet"
+    ) as wait_for_packet_mock, unittest.mock.patch("time.sleep") as sleep_mock:
         packet_queue = queue.Queue()
         packet_queue.put(None)
         packet_queue.put(None)
         packet_queue.put(packet)
         packet_queue.put(None)
         packet_queue.put(packet)
-        receive_packet_mock.side_effect = packet_queue.get
+        wait_for_packet_mock.side_effect = (
+            lambda gdo0_gpio_line_name, timeout_seconds: packet_queue.get()
+        )
+        assert next(measurement_iter) is None  # timeout or error fetching package
+        sleep_mock.assert_called_once_with(1)  # protect against package flood
+        assert next(measurement_iter) is None
         with unittest.mock.patch.object(
             sensor, "_parse_transmission", return_value="dummy"
         ) as parse_transmission_mock:
             assert next(measurement_iter) == "dummy"
+        assert sleep_mock.call_count == 3
         sensor._transceiver.__enter__.assert_called_once_with()  # pylint: disable=no-member; false positive
-        assert receive_packet_mock.call_count == 3
+        assert wait_for_packet_mock.call_count == 3
         assert parse_transmission_mock.call_count == 1
         (
             parse_transmission_args,
@@ -181,25 +182,30 @@ def test_receive():
             parse_transmission_args[0],
             numpy.array([255, 168] + [0] * 23, dtype=numpy.uint8),
         )
+        assert next(measurement_iter) is None
         with unittest.mock.patch.object(
             sensor, "_parse_transmission", return_value="dummy2"
         ):
             assert next(measurement_iter) == "dummy2"
+        assert sleep_mock.call_count == 5
     # pylint: disable=no-member; false positive
     sensor._transceiver.unlock_spi_device.assert_not_called()
 
 
 def test_receive_failed_to_decode(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
-        sensor = wireless_sensor.FT017TH()
-    measurement_iter = sensor.receive()
+        sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
+    measurement_iter = sensor.receive(timeout_seconds=21)
     with unittest.mock.patch.object(
-        sensor,
-        "_receive_packet",
+        sensor._transceiver,
+        "_wait_for_packet",
         return_value=cc1101._ReceivedPacket(
-            data=b"\0" * 23, rssi_index=0, checksum_valid=True, link_quality_indicator=0
+            payload=b"\0" * 23,
+            rssi_index=0,
+            checksum_valid=True,
+            link_quality_indicator=0,
         ),
-    ) as receive_packet_mock, unittest.mock.patch.object(
+    ) as wait_for_packet_mock, unittest.mock.patch.object(
         sensor,
         "_parse_transmission",
         side_effect=[
@@ -207,12 +213,17 @@ def test_receive_failed_to_decode(caplog):
             wireless_sensor.DecodeError("dummy error 1"),
             "dummy measurement",
         ],
-    ) as parse_transmission_mock, caplog.at_level(
+    ) as parse_transmission_mock, unittest.mock.patch(
+        "time.sleep"
+    ) as sleep_mock, caplog.at_level(
         logging.DEBUG
     ):
+        assert next(measurement_iter) is None
+        assert next(measurement_iter) is None
         assert next(measurement_iter) == "dummy measurement"
-    assert receive_packet_mock.call_count == 3
+    assert wait_for_packet_mock.call_count == 3
     assert parse_transmission_mock.call_count == 3
+    assert sleep_mock.call_count == 3
     decode_error_log_records = [
         (r.message, r.exc_info[1])
         for r in caplog.records
@@ -233,7 +244,7 @@ def test_receive_failed_to_decode(caplog):
 
 def test_receive_unexpected_packet_length(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
-        sensor = wireless_sensor.FT017TH()
+        sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement = wireless_sensor.Measurement(
         decoding_timestamp=datetime.datetime.now(),
         temperature_degrees_celsius=21.0,
@@ -247,8 +258,11 @@ def test_receive_unexpected_packet_length(caplog):
             wireless_sensor._UnexpectedPacketLengthError,
             measurement,
         ],
-    ), caplog.at_level(logging.INFO):
-        assert next(sensor.receive()) == measurement
+    ), unittest.mock.patch("time.sleep") as sleep_mock, caplog.at_level(logging.INFO):
+        measurement_iter = sensor.receive(timeout_seconds=21)
+        assert next(measurement_iter) is None
+        assert next(measurement_iter) is None
+        assert next(measurement_iter) == measurement
     assert (
         # pylint: disable=no-member; false positive
         sensor._transceiver.__enter__.call_count
@@ -266,16 +280,53 @@ def test_receive_unexpected_packet_length(caplog):
         ]
         * 2
     )
+    assert sleep_mock.call_count == 3
+
+
+@pytest.mark.parametrize("duration_seconds", (1, 21))
+def test__sleep_single(duration_seconds):
+    itr = wireless_sensor.FT017TH._sleep(
+        duration_seconds=duration_seconds, yield_interval=42
+    )
+    with unittest.mock.patch("time.sleep") as sleep_mock:
+        assert next(itr) is None
+        assert sleep_mock.call_count == 0
+        assert next(itr) is None
+        assert sleep_mock.call_count == 1
+        args, kwargs = sleep_mock.call_args
+        assert len(args) == 1
+        assert args[0] == pytest.approx(duration_seconds, abs=0.1)
+        assert not kwargs
+
+
+def test__sleep_multiple():
+    with unittest.mock.patch("time.sleep") as sleep_mock, unittest.mock.patch(
+        "time.time", side_effect=[100, 100.01, 103, 110, 113, 120, 121, 123]
+    ):
+        for idx, item in enumerate(
+            wireless_sensor.FT017TH._sleep(duration_seconds=23, yield_interval=7)
+        ):
+            assert item is None
+            if idx == 0:
+                assert sleep_mock.call_count == 0
+            else:
+                assert sleep_mock.call_count == 1
+                sleep_mock.assert_called_once_with(7)
+                sleep_mock.reset_mock()
+    assert idx + 1 == 3  # pylint: disable=undefined-loop-variable
+    assert sleep_mock.call_count == 1
+    sleep_mock.assert_called_once_with(2)
 
 
 def test_receive_locked(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
-        sensor = wireless_sensor.FT017TH()
+        sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement = wireless_sensor.Measurement(
         decoding_timestamp=datetime.datetime.now(),
         temperature_degrees_celsius=21.0,
         relative_humidity=42.0,
     )
+    measurement_iter = sensor.receive(timeout_seconds=21)
     with unittest.mock.patch.object(
         sensor._transceiver,
         "__enter__",
@@ -288,26 +339,49 @@ def test_receive_locked(caplog):
     ) as enter_mock, unittest.mock.patch.object(
         sensor, "_receive_measurement", return_value=measurement
     ), unittest.mock.patch(
-        "time.sleep"
+        "time.sleep",
     ) as sleep_mock, caplog.at_level(
         logging.INFO
     ):
-        assert next(sensor.receive()) == measurement
+        assert next(measurement_iter) is None  # 1st BlockingIOError
+        assert sleep_mock.call_count == 0
+        assert next(measurement_iter) is None  # 1st BlockingIOError
+        assert sleep_mock.call_count == 1
+        assert sleep_mock.call_args[0][0] == pytest.approx(2, abs=0.01)
+        sleep_mock.reset_mock()
+        with unittest.mock.patch("time.time", return_value=time.time() + 3):
+            assert next(measurement_iter) is None  # 2nd BlockingIOError
+        sleep_mock.assert_called_once_with(0)  # remaining lock wait of 1st error
+        sleep_mock.reset_mock()
+        with unittest.mock.patch("time.time", return_value=time.time() + 8):
+            assert next(measurement_iter) is None  # 3rd BlockingIOError
+        sleep_mock.assert_called_once_with(0)  # remaining lock wait of 2nd error
+        sleep_mock.reset_mock()
+        with unittest.mock.patch("time.time", return_value=time.time() + 17):
+            assert next(measurement_iter) is measurement
+        assert sleep_mock.call_count == 2
+        assert sleep_mock.call_args_list[0][0][0] == 0  # remaining lock wait
+        assert sleep_mock.call_args_list[1][0][0] == 1  # flood protection
+        sleep_mock.reset_mock()
+        with unittest.mock.patch.object(
+            sensor, "_receive_measurement", side_effect=BlockingIOError
+        ):
+            assert next(measurement_iter) is None
+        sleep_mock.assert_called_once_with(1)  # flood protection
     assert enter_mock.call_count == 4
-    assert sleep_mock.call_args_list == [unittest.mock.call(s) for s in [2, 4, 8]]
     assert caplog.record_tuples == [
         (
             "wireless_sensor",
             logging.INFO,
             "SPI device locked, waiting {} seconds".format(seconds),
         )
-        for seconds in [2, 4, 8]
+        for seconds in [2, 4, 8, 2]
     ]
 
 
 def test_receive_no_reconfiguring(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
-        sensor = wireless_sensor.FT017TH()
+        sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement = wireless_sensor.Measurement(
         decoding_timestamp=datetime.datetime.now(),
         temperature_degrees_celsius=21.0,
@@ -315,8 +389,8 @@ def test_receive_no_reconfiguring(caplog):
     )
     with unittest.mock.patch.object(
         sensor, "_receive_measurement", return_value=measurement
-    ), caplog.at_level(logging.INFO):
-        measurement_iter = sensor.receive()
+    ), unittest.mock.patch("time.sleep"), caplog.at_level(logging.INFO):
+        measurement_iter = sensor.receive(timeout_seconds=21)
         for _ in range(3):
             assert next(measurement_iter) == measurement
     assert (
@@ -329,7 +403,9 @@ def test_receive_no_reconfiguring(caplog):
 
 def test_receive_unlock_spi_device(caplog):
     with unittest.mock.patch("cc1101.CC1101") as transceiver_class_mock:
-        sensor = wireless_sensor.FT017TH(unlock_spi_device=True)
+        sensor = wireless_sensor.FT017TH(
+            gdo0_gpio_line_name=b"GPIO24", unlock_spi_device=True
+        )
     transceiver_class_mock.assert_called_once_with(lock_spi_device=True)
     measurement = wireless_sensor.Measurement(
         decoding_timestamp=datetime.datetime.now(),
@@ -338,8 +414,8 @@ def test_receive_unlock_spi_device(caplog):
     )
     with unittest.mock.patch.object(
         sensor, "_receive_measurement", return_value=measurement
-    ), caplog.at_level(logging.DEBUG):
-        measurement_iter = sensor.receive()
+    ), unittest.mock.patch("time.sleep"), caplog.at_level(logging.DEBUG):
+        measurement_iter = sensor.receive(timeout_seconds=21)
         for _ in range(3):
             assert next(measurement_iter) == measurement
     transceiver_class_mock().unlock_spi_device.assert_called_once_with()